requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
boto3==1.36.25
botocore==1.36.25
braindecode==0.8.1
captum==0.7.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
dirhash==0.5.0
docstring-inheritance==2.2.2
EasyProcess==1.1
einops==0.8.1
entrypoint2==1.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gdown==5.2.0
grpcio==1.70.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
lightning==2.5.0.post0
lightning-utilities==0.12.0
loguru==0.7.3
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mne==1.9.0
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
npy-append-array==0.9.16
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
-e git+https://github.com/guidogagl/physioex@8a40e9058495e124c9238fcc52d54b9683e6846e#egg=physioex
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
psg-utils==0.1.7
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pyEDFlib==0.1.39
Pygments==2.19.1
PyNaCl==1.5.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pytz==2025.1
pyunpack==0.3
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.2.1
quantus==0.5.3
rarfile==4.2
requests==2.32.3
s3transfer==0.11.2
scantree==0.0.4
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
skorch==1.1.0
soundfile==0.13.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2025.3.30
torch==2.6.0
torchinfo==1.8.0
torchmetrics==1.6.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wfdb==4.2.0
yarl==1.18.3
